ENH: refactor start slots.py
I am refactoring the internal coordinates file
into multiple files, similarly to the way it is
organised in pyGSM, so that we can compare them
and develop in sync.
